fixed sha256

+++ ft_ssl.h
@@ -6,7 +6,7 @@
 /*   By: bogoncha <bogoncha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/16 16:54:46 by bogoncha          #+#    #+#             */
-/*   Updated: 2019/07/25 16:09:31 by bogoncha         ###   ########.fr       */
+/*   Updated: 2019/07/25 16:13:59 by bogoncha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,11 +53,11 @@ typedef struct	s_fmd5
 typedef struct		s_fsha
 {
 	unsigned long	hash[8];
-	unsigned		s[2];
-	unsigned		ch;
-	unsigned		maj;
-	unsigned		tmp1;
-	unsigned		tmp2;
+	unsigned long	s[2];
+	unsigned long	ch;
+	unsigned long	maj;
+	unsigned long	tmp1;
+	unsigned long	tmp2;
 	int				len;
 	__uint128_t		bitlen;
 	int				round;
